requirements.txt
CairoSVG==2.5.2
cairocffi==1.3.0
certifi==2019.11.28
charset-normalizer==2.0.6
cffi==1.14.6
cssselect2==0.4.1
defusedxml==0.7.1
idna==2.8
json5==0.9.6
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
numpy
pandas==1.3.3
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.14.0
tabulate==0.8.9
tinycss2==1.1.1
urllib3==1.25.8
webencodings==0.5.1
